Docker Compose based development environment
Use Docker Compose to run services locally and develop on a standard PC
rather than the Raspberry Pi. See `Docker.md` for details.

Some things are not working yet and will be addressed at a later stage:
- `nabd` is not available (Relies on Pi-specific hardware), meaning all
services that depend on it will not work. The plan is to write a fake
`nabd` daemon to fill the gap.
- Some pages of the web interfaces are not working correctly (for
example the Upgrade section). They probably make assumptions on the
environment that are not met by the Docker containers. It should be
fixable.

An alternative requirements file is used, currently not including the
dependencies that are harder to get for x86 (Kaldi, NLU...). This will
need to get fixed at some point to get all services running.

Most of this PR are new files, the only practical change is to enable
Gunicorn to serve static files on nabweb when the development
environment is used (based on an environment variable), as it doesn't
use Nginx for simplicity sake.

This is just a starting point at this stage, but it already allows
hacking on the web interface.