Docker Compose based development environment
Use Docker Compose to run services locally and develop on a standard PC
rather than the Raspberry Pi via SSH, and avoiding the need for a test
and a prod rabbit. See `Docker.md` for details.

Some things are not working yet and will be addressed at a later stage:
- `nabd` is not available (Relies on Pi-specific hardware), meaning all
services that depend on it will not work. The plan is to write a fake
`nabd` daemon to fill the gap.
- Some pages of the web interfaces are not working correctly (for
example the Upgrade section). They probably make assumptions on the
environment that are not met by the Docker containers. It should be
fixable.

An alternative requirements file is used, currently not including the
dependencies that are harder to get for x86 (Kaldi ASR...). This will
need to get fixed at some point to get all services running.

Most of this PR are new files, the only practical changes are:
- Enable Gunicorn to serve static files on nabweb when the development
environment is used (based on an environment variable), as it doesn't
use Nginx for simplicity sake
- Small error handling change when the Raspberry Pi model cannot be read
from `/proc`.

This is just a starting point at this stage, but it already allows
hacking on the web interface.