Next step in Docker-based development environment
Provide a dummy nabd daemon that can be run on a regular computer
without the real hardware. This daemon runs in a separate Docker
container that all services can access.

The nabd connection host is changed from 127.0.0.1 to `nabdevd`
(container name = hostname) when running in development mode, based on an
environment variable. This allows to run all services, each in their own
container.

The dummy daemon is named `nabdevd` and lives in a separate repository:
https://gitlab.com/nguillaumin/nabdevd . It may make sense to move it
inside the nabaztag2018 GitHub account at some point.

The dummy daemon is limited for now, but implements enough of the
protocol for the services to run. The plan is to simulate as many
features as possible of the real nabd, possibly with some control
interface to simulate button presses or ears manipulation.